src/search_bot/requirements.txt
python-dotenv

python-telegram-bot
telegramify-markdown

langchain
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-openai
langchain-ollama
langchain-community

langsmith
langgraph

sqlalchemy